+++ vulnerabilities/pipelines/compute_package_risk.py
@@ -75,19 +75,32 @@ class ComputePackageRiskPipeline(VulnerableCodePipeline):
         )
 
     def compute_and_store_package_risk_score(self):
+        affected_packages = (
+            Package.objects.filter(affected_by_vulnerabilities__isnull=False).prefetch_related(
+                "affectedbypackagerelatedvulnerability_set__vulnerability",
+                "affectedbypackagerelatedvulnerability_set__vulnerability__references",
+                "affectedbypackagerelatedvulnerability_set__vulnerability__severities",
+                "affectedbypackagerelatedvulnerability_set__vulnerability__exploits",
+            )
+        ).distinct()
+
         affected_packages = Package.objects.filter(
             affected_by_vulnerabilities__isnull=False
         ).distinct()
 
         self.log(f"Calculating risk for {affected_packages.count():,d} affected package records")
 
-        progress = LoopProgress(total_iterations=affected_packages.count(), logger=self.log)
+        progress = LoopProgress(
+            total_iterations=affected_packages.count(),
+            logger=self.log,
+            progress_step=5,
+        )
 
         updatables = []
         updated_package_count = 0
-        batch_size = 5000
+        batch_size = 10000
 
-        for package in progress.iter(affected_packages.paginated()):
+        for package in progress.iter(affected_packages.paginated(per_page=batch_size)):
             risk_score = compute_package_risk(package)
 
             if not risk_score: